+++ cornerstone-web/src/views/login/index.vue
@@ -26,13 +26,19 @@
         @keyup.enter="handleLogin"
       >
         <el-form-item label="用户名" prop="username">
-          <el-input v-model="form.username" placeholder="请输入用户名" :prefix-icon="User" />
+          <el-input
+            v-model="form.username"
+            placeholder="请输入用户名"
+            autocomplete="username"
+            :prefix-icon="User"
+          />
         </el-form-item>
         <el-form-item label="密码" prop="password">
           <el-input
             v-model="form.password"
             type="password"
             placeholder="请输入密码"
+            autocomplete="current-password"
             show-password
             :prefix-icon="Lock"
           />